# docs/swagger.yaml
basePath: /api
definitions:
...
  consts.EditorType:
    enum:
    - 0
    - 1
    format: int32
    type: integer
    x-enum-varnames:
    - EditorTypeMarkdown
    - EditorTypeRichText
...
  consts.MFAType:
    enum:
    - 0
    - 1
    format: int32
    type: integer
    x-enum-varnames:
    - MFANone
    - MFATFATotp
  consts.PostStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-varnames:
    - PostStatusPublished
    - PostStatusDraft
    - PostStatusRecycle
    - PostStatusIntimate
...
  dto.BaseDTO:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  dto.CategoryDTO:
    properties:
      createTime:
        type: integer
      description:
        type: string
      fullPath:
        type: string
      id:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      password:
        type: string
      priority:
        type: integer
      slug:
        type: string
      thumbnail:
        type: string
      type:
        type: integer
    type: object
...
  dto.Meta:
    properties:
      createTime:
        type: integer
      id:
        type: integer
      key:
        type: string
      postId:
        type: integer
      value:
        type: string
    type: object
...
  dto.Page:
    properties:
      content: {}
      hasContent:
        type: boolean
      hasNext:
        type: boolean
      hasPrevious:
        type: boolean
      isEmpty:
        type: boolean
      isFirst:
        type: boolean
      isLast:
        type: boolean
      pageNum:
        type: integer
      pages:
        type: integer
      rpp:
        type: integer
      total:
        type: integer
    type: object
...
  dto.Post:
    properties:
      createTime:
        type: integer
      disallowComment:
        type: boolean
      editTime:
        type: integer
      editorType:
        $ref: '#/definitions/consts.EditorType'
      fullPath:
        type: string
      id:
        type: integer
      likes:
        type: integer
      metaDescription:
        type: string
      metaKeywords:
        type: string
      password:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/consts.PostStatus'
      summary:
        type: string
      template:
        type: string
      thumbnail:
        type: string
      title:
        type: string
      topPriority:
        type: integer
      topped:
        type: boolean
      updateTime:
        type: integer
      visits:
        type: integer
      wordCount:
        type: integer
    type: object
...
  dto.Statistic:
    properties:
      birthday:
        type: integer
      categoryCount:
        type: integer
      commentCount:
        type: integer
      establishDays:
        type: integer
      journalCount:
        type: integer
      likeCount:
        type: integer
      linkCount:
        type: integer
      postCount:
        type: integer
      tagCount:
        type: integer
      visitCount:
        type: integer
    type: object
  dto.StatisticWithUser:
    properties:
      birthday:
        type: integer
      categoryCount:
        type: integer
      commentCount:
        type: integer
      establishDays:
        type: integer
      journalCount:
        type: integer
      likeCount:
        type: integer
      linkCount:
        type: integer
      postCount:
        type: integer
      tagCount:
        type: integer
      user:
        $ref: '#/definitions/dto.User'
      visitCount:
        type: integer
    type: object
  dto.Tag:
    properties:
      color:
        type: string
      createTime:
        type: integer
      fullPath:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      thumbnail:
        type: string
    type: object
...
  dto.User:
    properties:
      avatar:
        type: string
      createTime:
        type: integer
      description:
        type: string
      email:
        type: string
      id:
        type: integer
      mfaType:
        $ref: '#/definitions/consts.MFAType'
      nickname:
        type: string
      updateTime:
        type: integer
      username:
        type: string
    type: object
...
  param.Meta:
    properties:
      key:
        type: string
      value:
        type: string
    required:
    - key
    - value
    type: object
...
  param.Post:
    properties:
      categoryIds:
        items:
          type: integer
        type: array
      content:
        type: string
      createTime:
        type: integer
      disallowComment:
        type: boolean
      editTime:
        type: integer
      editorType:
        $ref: '#/definitions/consts.EditorType'
      metaDescription:
        type: string
      metaKeywords:
        type: string
      metas:
        items:
          $ref: '#/definitions/param.Meta'
        type: array
      originalContent:
        type: string
      password:
        maxLength: 255
        type: string
      slug:
        maxLength: 255
        type: string
      status:
        allOf:
        - $ref: '#/definitions/consts.PostStatus'
        minimum: 0
      summary:
        type: string
      tagIds:
        items:
          type: integer
        type: array
      template:
        maxLength: 255
        type: string
      thumbnail:
        type: string
      title:
        maxLength: 100
        minLength: 1
        type: string
      topPriority:
        minimum: 0
        type: integer
      updateTime:
        type: integer
    type: object
...
  vo.PostDetailVO:
    properties:
      categories:
        items:
          $ref: '#/definitions/dto.CategoryDTO'
        type: array
      categoryIds:
        items:
          type: integer
        type: array
      commentCount:
        type: integer
      content:
        type: string
      createTime:
        type: integer
      disallowComment:
        type: boolean
      editTime:
        type: integer
      editorType:
        $ref: '#/definitions/consts.EditorType'
      fullPath:
        type: string
      id:
        type: integer
      likes:
        type: integer
      metaDescription:
        type: string
      metaIds:
        items:
          type: integer
        type: array
      metaKeywords:
        type: string
      metas:
        items:
          $ref: '#/definitions/dto.Meta'
        type: array
      originalContent:
        type: string
      password:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/consts.PostStatus'
      summary:
        type: string
      tagIds:
        items:
          type: integer
        type: array
      tags:
        items:
          $ref: '#/definitions/dto.Tag'
        type: array
      template:
        type: string
      thumbnail:
        type: string
      title:
        type: string
      topPriority:
        type: integer
      topped:
        type: boolean
      updateTime:
        type: integer
      visits:
        type: integer
      wordCount:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    AirPress 博客系统后端 API 文档,通过 swag 注释驱动自动生成。
    包含后台管理(/api/admin)与前台内容(/api/content)两组接口。
  title: AirPress API
  version: "1.0"
paths:
...
  /admin/posts:
    get:
      consumes:
      - application/json
      description: 支持按关键词、状态、分类、标签过滤,支持排序与分页
      parameters:
      - description: 页码(从0开始)
        example: 0
        in: query
        name: page
        type: integer
      - description: 每页数量
        example: 10
        in: query
        name: size
        type: integer
      - collectionFormat: multi
        description: 排序字段,如 createTime,desc
        in: query
        items:
          type: string
        name: sort
        type: array
      - description: 标题关键词
        in: query
        name: keyword
        type: string
      - collectionFormat: multi
        description: 状态过滤
        in: query
        items:
          type: integer
        name: statuses
        type: array
      - description: 分类ID
        in: query
        name: categoryId
        type: integer
      - description: 标签ID
        in: query
        name: tagId
        type: integer
      - description: true返回详情VO,false返回精简DTO
        in: query
        name: more
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/dto.Page'
                  - properties:
                      content:
                        items:
                          $ref: '#/definitions/dto.Post'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 分页查询文章列表
      tags:
      - Admin.Post
    post:
      consumes:
      - application/json
      description: 创建一篇新文章,返回创建后的详情
      parameters:
      - description: 文章参数
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/param.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  $ref: '#/definitions/vo.PostDetailVO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 创建文章
      tags:
      - Admin.Post
  /admin/posts/{postID}:
    delete:
      description: 根据文章ID删除指定文章
      parameters:
      - description: 文章ID
        example: 1
        in: path
        name: postID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 删除文章
      tags:
      - Admin.Post
    get:
      description: 返回文章详情,包含正文、评论数等
      parameters:
      - description: 文章ID
        example: 1
        in: path
        name: postID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  $ref: '#/definitions/vo.PostDetailVO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 根据文章ID获取详情
      tags:
      - Admin.Post
    put:
      consumes:
      - application/json
      description: 根据文章ID更新文章内容
      parameters:
      - description: 文章ID
        example: 1
        in: path
        name: postID
        required: true
        type: integer
      - description: 文章参数
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/param.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  $ref: '#/definitions/vo.PostDetailVO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 更新文章
      tags:
      - Admin.Post
...
  /admin/statistics:
    get:
      description: 返回文章数、评论数、分类数、访问量等统计信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto.Statistic'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 获取博客统计数据
      tags:
      - Admin.Statistic
  /admin/statistics/user:
    get:
      description: 返回博客统计信息,同时附带当前登录管理员信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.BaseDTO'
            - properties:
                data:
                  $ref: '#/definitions/dto.StatisticWithUser'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.BaseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.BaseDTO'
      security:
      - AdminApiKey: []
      summary: 获取含用户信息的统计数据
      tags:
      - Admin.Statistic
...
securityDefinitions:
  AdminApiKey:
    in: header
    name: Admin-Authorization
    type: apiKey
  ContentApiKey:
    in: header
    name: Content-Authorization
    type: apiKey
swagger: "2.0"